requirements.txt
requests==2.27.1
Django==3.1.4
django-crispy-forms==1.14.0
django-filter==21.1